Realizei o povoamento do banco de dados bem como a primeira migration. Exceções personalizadas, serviços de classes e algumas telas iniciais. Estou organizando o projeto de uma forma parecida com o MVC

+++ WebSales - WebForms e EntityFramework/Models/Department.cs
@@ -53,7 +53,7 @@ namespace WebSales___WebForms_e_EntityFramework.Models
         }
 
         // Retorna a soma de todas as vendas dos vendedores
-        public float TotalSales()
+        public double TotalSales()
         {
             return Sellers.Select(s => s.TotalSales()).Sum();
         }
